+++ src/static/js/mugshot.js
@@ -1,6 +1,11 @@
-
 (function($) {
-
+	var MSG_INTRO = 'To use this app you need to grant permission to use the webcam from the top right of the screen!';
+	var MSG_NICE_SHOT = 'Nice Shot!<br />Click Save Photo to upload to the server!';
+	var MSG_SAVED = 'Your mugshot has been saved!';
+	var MSG_NOT_SUPPORTED = 'Your browser is not compatible with this app.<br />Try <a href="http://chrome.google.com" target="_BLANK">Chrome</a> or <a href="http://www.firefox.com" target="_BLANK">Firefox</a>';
+	
+	var ERR_PERMISSION_DENIED = 1;
+	
 	var streaming = false;
 	var video = document.querySelector('#video');
 	var canvas = document.querySelector('#canvas');
@@ -11,8 +16,6 @@
 	
 	var width = 300;
 	var height = 0;
-
-	navigator.getMedia = navigator.getUserMedia || navigator.webkitGetUserMedia || navigator.mozGetUserMedia || navigator.msGetUserMedia;
 
 	var constraints = {
 		video: true,
@@ -30,15 +33,13 @@
 	}
 	
 	function errorCallback(err) {
-		var PERMISSION_DENIED = 1;
-		if (err.code == PERMISSION_DENIED) {
+		if (err.code == ERR_PERMISSION_DENIED) {
 			alert('You need to grant this app permission to use your webcam!');
 		}
 		console.log("An error occured! " + err);
 	}
 	
 	video.addEventListener('canplay', function(ev) {
-		console.log('now')
 		if (!streaming) {
 			height = video.videoHeight / (video.videoWidth/width);
 			video.setAttribute('width', width);
@@ -59,8 +60,7 @@
 		$('#video').hide();
 		$('#photo').show();
 
-		$('#intro').hide();
-		$('#niceShot').show();
+		$('#msg').html(MSG_NICE_SHOT);
 
 		$('#btnCapture').hide();
 		$('#btnRetake').show();
@@ -68,6 +68,12 @@
 	}
 	
 	function savePicture() {
+		if (!EMAIL) {
+			if (confirm('You have to log in to save Mug Shots!\n\nLog in now?!'))
+				location.href = LOGIN_URL;
+			return;
+		}
+		
 		var data = canvas.toDataURL('image/png');
 		$.ajax({
 			type: 'POST',
@@ -75,13 +81,12 @@
 			data: { imgBase64: data }
 		}).done(function(id) {
 			console.log('saved with id ' + id);
-			$('#mugshots').prepend($('<img />').attr('src', '/photo/'+id).attr('id', 'mugshot_'+id).addClass('mugshot').click(function() { deletePicture(id); }));
+			$('#mugshots').prepend($('<img />').attr('src', '/photo/'+id).attr('id', 'mugshot_'+id).data('id', id).data('email', EMAIL).addClass('mugshot').click(function() { deletePicture(id); }));
 			
 			$('#video').show();
 			$('#photo').hide();
 			
-			$('#niceShot').hide();
-			$('#saved').show();
+			$('#msg').html(MSG_SAVED);
 			
 			$('#btnRetake').hide();
 			$('#btnSave').hide();
@@ -90,13 +95,24 @@
 	}
 	
 	function deletePicture(id) {
-		$.ajax({
-			type: 'DELETE',
-			url: '/photo/' + id,
-		}).done(function() {
-			console.log('deleted with id ' + id);
-			$('#mugshot_' + id).remove();
-		});
+		if (!EMAIL)
+			return;
+		
+		var email = $('#mugshot_'+id).data('email');
+		if (email != EMAIL) {
+			alert('You can only delete your own Mug Shots!');
+			return;
+		}
+		
+		if (confirm('Delete this Mug Shot?!')) {
+			$.ajax({
+				type: 'DELETE',
+				url: '/photo/' + id,
+			}).done(function() {
+				console.log('deleted with id ' + id);
+				$('#mugshot_' + id).remove();
+			});
+		}
 	}
 
 	$('#btnCapture').click(function() {
@@ -111,8 +127,6 @@
 		$('#video').show();
 		$('#photo').hide();
 
-		$('#niceShot').hide();
-
 		$('#btnCapture').show();
 		$('#btnRetake').hide();
 		$('#btnSave').hide();
@@ -122,7 +136,21 @@
 		var id = $(this).data('id');
 		deletePicture(id);
 	});
+
+	function init() {
+		navigator.getMedia = navigator.getUserMedia || navigator.webkitGetUserMedia || navigator.mozGetUserMedia || navigator.msGetUserMedia;
+		if (navigator.getMedia) {
+			navigator.getMedia(constraints, successCallback, errorCallback);
+			$('#msg').html(MSG_INTRO);
+		} else {
+			$('#msg').html(MSG_NOT_SUPPORTED).addClass('error');
+			$('#btnCapture').hide();
+			alert('Your browser is not compatible with this app.\n\nTry Chrome -> chrome.google.com');
+		}
+	}
 	
-	navigator.getMedia(constraints, successCallback, errorCallback);
+	$(function() {
+		init();
+	});
 	
 })(jQuery);
